projects: ltc2983: use the example infrastructure
Migrate ltc2983 to the shared examples infrastructure (examples.mk):
replace the four Makefile example toggles (BASIC/IIO/ADT7604_BASIC/
ADT7604_IIO) with EXAMPLE := iio + include examples.mk, give each example
its own src/examples/<name>/example.mk (carrying the -D<EXAMPLE> define
that common_data.{c,h} use to select the sensor set, plus IIOD=y and the
iio_ltc2983 sources for the iio variants), drop examples_src.mk and
platform_includes.h, and shorten the iio example directories
(iio_example -> iio, adt7604_iio_example -> adt7604_iio). main.c now
defers to example_main(); each basic example owns its UART stdio init and
each iio example includes iio_app.h directly. Update the README to select
the example with make EXAMPLE=<name> on the command line.
