+++ projects/ltc2983/src/examples/adt7604_basic/adt7604_basic_example.c
@@ -30,9 +30,9 @@
  * EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  *******************************************************************************/
 
-#include "adt7604_basic_example.h"
 #include "common_data.h"
 #include "ltc2983.h"
+#include "no_os_uart.h"
 #include "no_os_delay.h"
 #include "no_os_print_log.h"
 
@@ -42,14 +42,21 @@
  * @return ret - Result of the example execution. If working correctly, will
  *               execute continuously the while(1) loop and will not return.
  *******************************************************************************/
-int adt7604_basic_example_main()
+int example_main()
 {
 	struct ltc2983_desc *dev;
+	struct no_os_uart_desc *uart;
 	uint32_t raw_cov;
 	uint8_t type;
 	int val;
 	int ret;
 	int i;
+
+	ret = no_os_uart_init(&uart, &uip);
+	if (ret)
+		return ret;
+
+	no_os_uart_stdio(uart);
 
 	pr_info("\r\nRunning EVAL-ADT7604-AZ Basic Example\r\n");
