+++ projects/ltc2983/src/examples/adt7604_basic/example.mk
@@ -1,0 +1,1 @@
+CFLAGS += -DADT7604_BASIC_EXAMPLE=1
