+++ projects/ltc2983/src/examples/basic/example.mk
@@ -1,0 +1,1 @@
+CFLAGS += -DBASIC_EXAMPLE=1
