--- projects/ltc2983/src/examples/examples_src.mk
@@ -1,34 +1,0 @@
-ifeq (y,$(strip $(IIO_EXAMPLE)))
-IIOD=y
-CFLAGS += -DIIO_EXAMPLE=1
-SRCS += $(PROJECT)/src/examples/iio_example/iio_example.c
-INCS += $(PROJECT)/src/examples/iio_example/iio_example.h
-endif
-
-ifeq (y,$(strip $(BASIC_EXAMPLE)))
-CFLAGS += -DBASIC_EXAMPLE=1
-SRCS += $(PROJECT)/src/examples/basic/basic_example.c
-INCS += $(PROJECT)/src/examples/basic/basic_example.h
-endif
-
-ifeq (y,$(strip $(ADT7604_BASIC_EXAMPLE)))
-CFLAGS += -DADT7604_BASIC_EXAMPLE=1
-SRCS += $(PROJECT)/src/examples/adt7604_basic/adt7604_basic_example.c
-INCS += $(PROJECT)/src/examples/adt7604_basic/adt7604_basic_example.h
-endif
-
-ifeq (y,$(strip $(ADT7604_IIO_EXAMPLE)))
-IIOD=y
-CFLAGS += -DADT7604_IIO_EXAMPLE=1
-SRCS += $(PROJECT)/src/examples/adt7604_iio_example/adt7604_iio_example.c
-INCS += $(PROJECT)/src/examples/adt7604_iio_example/adt7604_iio_example.h
-endif
-
-ifeq (y,$(strip $(IIOD)))
-SRC_DIRS += $(NO-OS)/iio/iio_app
-INCS += $(DRIVERS)/temperature/ltc2983/iio_ltc2983.h
-SRCS += $(DRIVERS)/temperature/ltc2983/iio_ltc2983.c
-
-INCS += $(INCLUDE)/no_os_list.h \
-		$(PLATFORM_DRIVERS)/$(PLATFORM)_uart.h
-endif
