--- projects/ltc2983/src/examples/iio_example/iio_example.c
@@ -1,81 +1,0 @@
-/********************************************************************************
- *   @file   iio_example.c
- *   @brief  IIO example code for the ltc2983 project
- *   @author John Erasmus Mari Geronimo (johnerasmusmari.geronimo@analog.com)
- ********************************************************************************
- * Copyright 2024(c) Analog Devices, Inc.
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are met:
- *
- * 1. Redistributions of source code must retain the above copyright notice,
- *    this list of conditions and the following disclaimer.
- *
- * 2. Redistributions in binary form must reproduce the above copyright notice,
- *    this list of conditions and the following disclaimer in the documentation
- *    and/or other materials provided with the distribution.
- *
- * 3. Neither the name of Analog Devices, Inc. nor the names of its
- *    contributors may be used to endorse or promote products derived from this
- *    software without specific prior written permission.
- *
- * THIS SOFTWARE IS PROVIDED BY ANALOG DEVICES, INC. “AS IS” AND ANY EXPRESS OR
- * IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF
- * MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO
- * EVENT SHALL ANALOG DEVICES, INC. BE LIABLE FOR ANY DIRECT, INDIRECT,
- * INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
- * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA,
- * OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF
- * LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
- * NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
- * EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- *******************************************************************************/
-
-#include "iio_example.h"
-#include "common_data.h"
-#include "no_os_print_log.h"
-
-/*******************************************************************************
- * @brief IIO example main execution.
- *
- * @return ret - Result of the example execution. If working correctly, will
- *               execute continuously function iio_app_run and will not return.
- *******************************************************************************/
-int iio_example_main()
-{
-	int ret;
-	struct ltc2983_iio_desc *ltc2983_iio_dev;
-	struct ltc2983_iio_desc_init_param ltc2983_iio_ip;
-	struct iio_app_desc *app;
-	struct iio_app_init_param app_init_param = {0};
-
-	ltc2983_iio_ip.ltc2983_desc_init_param = &ltc2983_ip;
-	ret = ltc2983_iio_init(&ltc2983_iio_dev, &ltc2983_iio_ip);
-	if (ret)
-		return ret;
-
-	struct iio_app_device iio_devices[] = {
-		{
-			.name = "ltc2983",
-			.dev = ltc2983_iio_dev,
-			.dev_descriptor = ltc2983_iio_dev->iio_dev,
-		},
-	};
-
-	app_init_param.devices = iio_devices;
-	app_init_param.nb_devices = NO_OS_ARRAY_SIZE(iio_devices);
-	app_init_param.uart_init_params = uip;
-
-	ret = iio_app_init(&app, app_init_param);
-	if (ret)
-		goto error;
-
-	ret = iio_app_run(app);
-	if (ret)
-		pr_info("Error: iio_app_run: %d\r\n", ret);
-
-	iio_app_remove(app);
-error:
-	ltc2983_iio_remove(ltc2983_iio_dev);
-	return ret;
-}
